+++ xradar/io/backends/__init__.py
@@ -15,6 +15,7 @@ Xarray Backends
 """
 
 from .cfradial1 import *  # noqa
+from .gamic import *  # noqa
 from .odim import *  # noqa
 
 __all__ = [s for s in dir() if not s.startswith("_")]