+++ xradar/io/backends/__init__.py
@@ -11,10 +11,13 @@ Xarray Backends
 
 .. automodule:: xradar.io.backends.cfradial1
 .. automodule:: xradar.io.backends.odim
+.. automodule:: xradar.io.backends.furuno
+
 
 """
 
 from .cfradial1 import *  # noqa
+from .furuno import *  # noqa
 from .odim import *  # noqa
 
 __all__ = [s for s in dir() if not s.startswith("_")]