+++ docs/history.md
@@ -2,6 +2,10 @@
 
 ## 0.7.0 (2022-09-21)
 
+* add Furuno SCN/SCNX importer ({pull}`30`) by [@kmuehlbauer](https://github.com/kmuehlbauer)
+
+## 0.7.0 (2022-09-21)
+
 * Add zenodo badges to README.md ({pull}`22`) by [@mgrover1](https://github.com/mgrover1)
 * Fix version on RTD ({pull}`23`) by [@kmuehlbauer](https://github.com/kmuehlbauer)
 * Add minimal documentation for Datamodel ({pull}`24`) by [@kmuehlbauer](https://github.com/kmuehlbauer)